# frontend/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "业务监控系统 API",
        "description": "业务监控系统后端接口文档，包含服务器资产管理、工作流编排、告警配置等功能模块。",
        "contact": {
            "name": "Monitor Team",
            "email": "monitor@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "开发环境"
        }
    ],
...
    "paths": {
...
        "/api/monitor-template": {
            "get": {
                "tags": [
                    "监控模板"
                ],
                "summary": "获取所有模板",
                "operationId": "getAll_1",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListMonitorTemplate"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "监控模板"
                ],
                "summary": "更新模板",
                "operationId": "update_6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MonitorTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "监控模板"
                ],
                "summary": "添加模板",
                "operationId": "add_5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MonitorTemplateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/monitor-template/{id}": {
            "get": {
                "tags": [
                    "监控模板"
                ],
                "summary": "获取模板详情",
                "operationId": "getById_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "模板ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMonitorTemplate"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "监控模板"
                ],
                "summary": "删除模板",
                "operationId": "delete_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "模板ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/monitor-template/category/{category}": {
            "get": {
                "tags": [
                    "监控模板"
                ],
                "summary": "按分类获取模板",
                "operationId": "getByCategory",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "分类名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListMonitorTemplate"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/alarm/template/{id}": {
            "get": {
                "tags": [
                    "告警管理"
                ],
                "summary": "获取告警模板详情",
                "operationId": "getTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAlarmTemplate"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "告警管理"
                ],
                "summary": "删除告警模板",
                "operationId": "deleteTemplate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
            "MonitorTemplateRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID，更新时必填",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "模板名称"
                    },
                    "category": {
                        "type": "string",
                        "description": "分类"
                    },
                    "collectScript": {
                        "type": "string",
                        "description": "采集脚本"
                    },
                    "defaultThreshold": {
                        "type": "string",
                        "description": "默认阈值规则（JSON格式）"
                    },
                    "defaultCron": {
                        "type": "string",
                        "description": "默认Cron表达式"
                    },
                    "alarmTemplateId": {
                        "type": "integer",
                        "description": "告警模板ID",
                        "format": "int64"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    }
                },
                "description": "监控模板请求"
            },
...
            "ApiResponseListMonitorTemplate": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/MonitorTemplate"
                        }
                    }
                },
                "description": "统一API响应"
            },
            "MonitorTemplate": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "模板名称"
                    },
                    "category": {
                        "type": "string",
                        "description": "分类（BASIC/COMPONENT/APPLICATION）"
                    },
                    "subCategory": {
                        "type": "string",
                        "description": "子分类（CPU/MySQL/Redis等）"
                    },
                    "collectType": {
                        "type": "string",
                        "description": "采集类型（SSH_SCRIPT/API/SQL）"
                    },
                    "collectScript": {
                        "type": "string",
                        "description": "采集脚本"
                    },
                    "paramSchema": {
                        "type": "string",
                        "description": "参数定义（JSON格式）"
                    },
                    "defaultThreshold": {
                        "type": "string",
                        "description": "默认告警阈值（JSON格式）"
                    },
                    "defaultCron": {
                        "type": "string",
                        "description": "默认Cron表达式"
                    },
                    "alarmTemplateId": {
                        "type": "integer",
                        "description": "关联的告警模板ID",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string",
                        "description": "模板描述"
                    },
                    "isSystem": {
                        "type": "integer",
                        "description": "是否为系统内置模板（1=是，0=否）",
                        "format": "int32"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "description": "排序顺序",
                        "format": "int32"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    },
                    "alarmTemplateName": {
                        "type": "string"
                    }
                },
                "description": "监控模板"
            },
            "ApiResponseMonitorTemplate": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "$ref": "#/components/schemas/MonitorTemplate"
                    }
                },
                "description": "统一API响应"
            },
...
            "AlarmTemplate": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "模板名称"
                    },
                    "subject": {
                        "type": "string",
                        "description": "邮件主题"
                    },
                    "content": {
                        "type": "string",
                        "description": "模板内容（支持变量替换）"
                    },
                    "contentType": {
                        "type": "string",
                        "description": "内容格式（TEXT/MARKDOWN）"
                    },
                    "isDefault": {
                        "type": "integer",
                        "description": "是否为默认模板（1=默认，0=非默认）",
                        "format": "int32"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    }
                },
                "description": "告警模板"
            },
...
            "ApiResponseAlarmTemplate": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AlarmTemplate"
                    }
                },
                "description": "统一API响应"
            },
...
        }
    }
}